+++ uTPro/Common/uTPro.Common/Helpers/SchemeUrlHelper.cs
@@ -1,14 +1,13 @@
-using Microsoft.AspNetCore.Http;
-
 namespace uTPro.Extension
 {
-    public static class SchemeUrlExtensions
+    public static class SchemeUrlHelper
     {
+        private const string PlaceholderHost = "invalid.utpro.local";
         public static string AddScheme(string urlRedirect, string schemeDefault = "https")
         {
             if (urlRedirect.StartsWith("/"))
             {
-                return schemeDefault + "://utpro.local" + urlRedirect;
+                return schemeDefault + $"://{PlaceholderHost}" + urlRedirect;
             }
             if (urlRedirect.StartsWith("http://") || urlRedirect.StartsWith("https://"))
             {
