+++ uTPro/Project/uTPro.Project.Web/Views/_ViewImports.cshtml
@@ -2,6 +2,7 @@
 
 @using uTPro.Extension;
 @using uTPro.Extension.CurrentSite;
+@using uTPro.Extension.UrlRouting;
 
 @using Umbraco.Extensions
 @using Umbraco.Cms.Web.Common.PublishedModels
