+++ uTPro/Project/uTPro.Project.Web/Views/uTPro/blockgrid/Components/PageList.cshtml
@@ -38,7 +38,7 @@
     var isCategoryFiltered = false;
     string? categoryFilterName = null;
     if ((filterCategory == null || !filterCategory.Any())
-        && Context.Items.TryGetValue(uTPro.Project.Web.Configure.CategoryLandingContentFinder.CategoryLandingItemKey, out var categoryLandingItem))
+        && Context.Items.TryGetValue(CategoryLandingContentFinder.CategoryLandingItemKey, out var categoryLandingItem))
     {
         var urlCategoryKey = (Guid)categoryLandingItem!;
         if (urlCategoryKey != Guid.Empty)
