+++ uTPro/Project/uTPro.Project.Web/Views/uTPro/blockgrid/Components/SearchResults.cshtml
@@ -134,7 +134,7 @@
     // ── Auto-filter search results by category from URL ──
     var isCategoryFiltered = false;
     string? categoryFilterName = null;
-    if (Context.Items.TryGetValue(uTPro.Project.Web.Configure.CategoryLandingContentFinder.CategoryLandingItemKey, out var categoryLandingItem))
+    if (Context.Items.TryGetValue(CategoryLandingContentFinder.CategoryLandingItemKey, out var categoryLandingItem))
     {
         var urlCategoryKey = (Guid)categoryLandingItem!;
         if (urlCategoryKey != Guid.Empty && results.Items.Any())
