+++ uTPro/Project/uTPro.Project.Web/uTPro.Project.Web.csproj
@@ -6,7 +6,7 @@
     <!-- SDK build/publish-time Brotli+GZip pre-compression of static web assets (.NET 9+). -->
     <CompressionEnabled>true</CompressionEnabled>
     <ErrorOnDuplicatePublishOutputFiles>false</ErrorOnDuplicatePublishOutputFiles>
-    <Version>17.9.0</Version>
+    <Version>17.10.0</Version>
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="Controllers\Api\**" />
